moved TabCompleteSellhead to tabcomplete package
