+++ src/main/java/io/github/thatsmusic99/headsplus/HeadsPlus.java
@@ -15,6 +15,7 @@ import io.github.thatsmusic99.headsplus.config.customheads.ConfigCustomHeads;
 import io.github.thatsmusic99.headsplus.inventories.InventoryManager;
 import io.github.thatsmusic99.headsplus.listeners.*;
 import io.github.thatsmusic99.headsplus.listeners.tabcompleting.TabComplete;
+import io.github.thatsmusic99.headsplus.listeners.tabcompleting.TabCompleteSellhead;
 import io.github.thatsmusic99.headsplus.managers.EntityDataManager;
 import io.github.thatsmusic99.headsplus.managers.PersistenceManager;
 import io.github.thatsmusic99.headsplus.storage.Favourites;
